remove/requirements.txt
Flask==2.2.3
torch==2.0.0
torchvision==0.15.0
transformers==4.35.0
Pillow==9.4.0